pyproject.toml
[project]
name = "exercise-rx-ai-apigateway"
version = "0.1.0"
description = "API网关服务"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.95.0",
    "httpx>=0.24.0",
    "uvicorn>=0.22.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
]
